#28 check balance

+++ front/front_push.rb
@@ -32,6 +32,7 @@ end
 post '/push' do
   token = the_human.tokens.find(params[:token])
   raise Baza::Urror, 'The token is inactive' unless token.active?
+  raise Baza::Urror, 'The balance is negative' unless the_human.account.balance.positive?
   Tempfile.open do |f|
     FileUtils.copy(params[:factbase][:tempfile], f.path)
     File.delete(params[:factbase][:tempfile])
